un-sandbox Markdown block
