+++ package.json
@@ -238,6 +238,7 @@
     "@githubnext/utils": "^0.15.4",
     "@githubocto/flat-ui": "^0.13.4",
     "@loadable/component": "^5.15.0",
+    "@mdx-js/runtime": "^1.6.22",
     "@octokit/rest": "^18.12.0",
     "@primer/components": "^31.1.0",
     "@react-three/drei": "^7.20.5",
@@ -273,6 +274,7 @@
     "react-syntax-highlighter": "^15.4.5",
     "react-use": "^17.3.1",
     "recharts": "^2.1.6",
+    "styled-components": "^5.3.3",
     "three": "^0.134.0"
   },
   "resolutions": {
@@ -282,6 +284,7 @@
   "devDependencies": {
     "@types/git-url-parse": "^9.0.1",
     "@types/lodash": "^4.14.176",
+    "@types/lz-string": "^1.3.34",
     "@types/parse-github-url": "^1.0.0",
     "@types/react": "^17.0.0",
     "@types/react-dom": "^17.0.0",
